+++ export_academy/models.py
@@ -13,13 +13,21 @@ from wagtail.admin.edit_handlers import (
 from wagtail.core.fields import RichTextField, StreamField
 from wagtail.core.models import Page
 
-from core.blocks import ButtonBlock
+from core.blocks import ButtonBlock, SingleRichTextBlock, TopicPageCardBlockRichText
 from core.constants import RICHTEXT_FEATURES__REDUCED
+from core.fields import single_struct_block_stream_field_factory
 from core.models import TimeStampedModel
+from export_academy import managers
 from export_academy.cms_panels import ExportAcademyPagePanels
 
 
 class Event(TimeStampedModel, ClusterableModel):
+    """
+    Represents an Export Academy event.
+
+    Includes Wagtail-specific types to enable Event objects to be managed from Wagtail admin.
+    """
+
     id = models.UUIDField(primary_key=True, default=uuid.uuid4)
     name = models.CharField(max_length=255)
     description = models.CharField(max_length=1000)
@@ -54,6 +62,8 @@ class Event(TimeStampedModel, ClusterableModel):
         ]
     )
 
+    objects = managers.EventQuerySet.as_manager()
+
     class Meta:
         ordering = ('-created', '-modified')
 
@@ -62,6 +72,12 @@ class Event(TimeStampedModel, ClusterableModel):
 
 
 class Registration(TimeStampedModel):
+    """
+    Represents an onboarding to Export Academy.
+
+    Captures data submitted via the Export Academy registration form.
+    """
+
     email = models.EmailField(primary_key=True)
     first_name = models.CharField(max_length=30)
     last_name = models.CharField(max_length=30)
@@ -75,6 +91,12 @@ class Registration(TimeStampedModel):
 
 
 class Booking(TimeStampedModel):
+    """
+    Represents the booking of an Event object.
+
+    Maps an Event object to a Registration object and registers a status.
+    """
+
     CONFIRMED = 'Confirmed'
     CANCELLED = 'Cancelled'
     STATUSES = (
@@ -91,9 +113,58 @@ class Booking(TimeStampedModel):
 class ExportAcademyHomePage(ExportAcademyPagePanels, Page):
     template = 'export_academy/landing_page.html'
 
+    hero_image = models.ForeignKey(
+        'core.AltTextImage',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
     hero_text = RichTextField(
         features=RICHTEXT_FEATURES__REDUCED,
         null=True,
         blank=True,
     )
     hero_cta = StreamField([('button', ButtonBlock(icon='cog'))], null=True, blank=True)
+
+    banner_label = models.CharField(
+        null=True,
+        blank=True,
+        max_length=255,
+        verbose_name='Banner label',
+    )
+
+    banner_content = models.CharField(
+        null=True,
+        blank=True,
+        max_length=255,
+        verbose_name='Banner Content',
+    )
+
+    steps_heading = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+        null=True,
+        blank=True,
+    )
+
+    steps = single_struct_block_stream_field_factory(
+        field_name='panel',
+        block_class_instance=SingleRichTextBlock(),
+        null=True,
+        blank=True,
+    )
+
+    panel_description = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+        null=True,
+        blank=True,
+    )
+
+    panels = single_struct_block_stream_field_factory(
+        field_name='panel',
+        block_class_instance=TopicPageCardBlockRichText(),
+        null=True,
+        blank=True,
+    )
+
+    next_cta = StreamField([('button', ButtonBlock())], null=True, blank=True)