+++ export_academy/views.py
@@ -1,13 +1,16 @@
 from django.urls import reverse_lazy
 from django.views.generic import FormView, ListView, TemplateView, UpdateView
+from django_filters.views import FilterView
 
 from config import settings
-from export_academy import forms, models
+from export_academy import filters, forms, models
 from export_academy.mixins import SaveAndSendNotifyMixin
 
 
-class EventListView(ListView):
+class EventListView(FilterView, ListView):
     model = models.Event
+    filterset_class = filters.EventFilter
+    template_name = 'export_academy/event_list.html'
 
     def get_context_data(self, **kwargs):
         ctx = super().get_context_data(**kwargs)