+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Enhancements
 * KLS-368 - Add tests for reskin fixes
 * KLS-395 - Update django to 3.2.18, directory-sso-api-client to >= 7.1.1
+* KLS-343 Add Export academy landing page
 
 ## [2.25.1](https://github.com/uktrade/great-cms/releases/tag/2.25.1)