+++ CHANGELOG.md
@@ -21,6 +21,8 @@ KLS-368 - Removed custom wagtailadmin JS and HTML causing page chooser issues
 
 ## [2.25.0](https://github.com/uktrade/great-cms/releases/tag/2.25.0)
 ### Bugs fixed
+* IOO-400 - International Online Offer using govuk frontend toolkit added styling to start page
+* IOO-402 - International Online Offer triage question 1 allowing user to select business sector
 * Update MonthYearInput test to get current year dynamically
 * KLS-370 - Update createPageChooser function to work with new wagtail parameters
 * KLS-80 - Correct form destinations and CTA style for Scottish offices
@@ -66,9 +68,6 @@ KLS-368 - Removed custom wagtailadmin JS and HTML causing page chooser issues
 ### Enhancements
 * KLS-266 - Get survey details from directory API
 * KLS-213 - Add phone number field to trade office contact forms
-
-* IOO-400 - International Online Offer using govuk frontend toolkit added styling to start page
-* IOO-402 - International Online Offer triage question 1 allowing user to select business sector
 
 ## [2.20.0](https://github.com/uktrade/great-cms/releases/tag/2.20.0)