+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Pre release
+### Enhancements
+* IOO-402 - International Online Offer triage question 1 allowing user to select business sector
 
 ## [2.26.0](https://github.com/uktrade/great-cms/releases/tag/2.26.0)
 ### Enhancements
@@ -19,6 +21,7 @@ KLS-368 - Removed custom wagtailadmin JS and HTML causing page chooser issues
 
 ### Enhancements
 * Updated UI component styling for reskin
+* IOO-400 - International Online Offer using govuk frontend toolkit added styling to start page
 
 ## [2.25.0](https://github.com/uktrade/great-cms/releases/tag/2.25.0)
 ### Bugs fixed
@@ -35,6 +38,7 @@ KLS-368 - Removed custom wagtailadmin JS and HTML causing page chooser issues
 ### Enhancements
 * KLS-213 - Add phone number field to trade office contact forms
 * KLS-266 - Get survey details from directory API
+
 
 
 ## [2.24.0](https://github.com/uktrade/great-cms/releases/tag/2.24.0)
@@ -67,8 +71,6 @@ KLS-368 - Removed custom wagtailadmin JS and HTML causing page chooser issues
 ### Enhancements
 * KLS-266 - Get survey details from directory API
 * KLS-213 - Add phone number field to trade office contact forms
-
-* IOO-400 - International Online Offer using govuk frontend toolkit added styling to start page
 
 ## [2.20.0](https://github.com/uktrade/great-cms/releases/tag/2.20.0)